+++ packages/shared/matches/extendMatchResult/extendKillStats.ts
@@ -1,3 +1,5 @@
+import { MAP_NAMES } from '../../riot-api/gameConstants';
+
 const thirtySeconds = 30 * 1000;
 const fifteenMinutes = 15 * 60 * 1000;
 const thirtyMinutes = 30 * 60 * 1000;
@@ -19,15 +21,16 @@ export default function extendKillStats(extendedMatchResult) {
 
   // add respawn time to kill/death
   // formula on summonersRift: 8 + 2.5 * level + [between 15-30] TIME-15min / 30 * 0.00425  + [between 30-45] TIME-30min / 30 * 0.003
+  // https://leagueoflegends.fandom.com/wiki/Death
   if (extendedMatchResult.isSummonersRift && extendedMatchResult.isModeClassic) {
     events.allKills.forEach(kill => {
-      let basetimer = 0;
+      let baseTimer = 0;
       if (kill.victimLevel <= 6) {
-        basetimer = 4 + 2 * kill.victimLevel;
+        baseTimer = 4 + 2 * kill.victimLevel;
       } else if (kill.victimLevel === 7) {
-        basetimer = 21;
+        baseTimer = 21;
       } else {
-        basetimer = 8 + 2.5 * kill.victimLevel;
+        baseTimer = 8 + 2.5 * kill.victimLevel;
       }
       let timeMultiplier = 1.0;
       if (kill.timestamp > fifteenMinutes && kill.timestamp <= thirtyMinutes) {
@@ -39,7 +42,12 @@ export default function extendKillStats(extendedMatchResult) {
       }
       // time multiplier maxes out at 50%
       timeMultiplier = Math.max(timeMultiplier, 1.5);
-      kill.deathTime = basetimer * timeMultiplier * 1000;
+      kill.deathTime = baseTimer * timeMultiplier * 1000;
+    });
+  } else if (extendedMatchResult.mapId === MAP_NAMES.HOWLING_ABYSS) {
+    // Level × 2 + 4 = Respawn time
+    events.allKills.forEach(kill => {
+      kill.deathTime = (kill.victimLevel * 2 + 4) * 1000;
     });
   }
 }